requirements.txt
icrawler==0.6.7
Pillow==10.1.0
PyQt5==5.15.10
PyQt5_sip==12.12.2
